requirements.txt
PyQt5

tensorflow

numpy

matplotlib

Pillow

scikit-learn